+++ tellurium/dev/deprecated/README.md
@@ -1,0 +1,5 @@
+# deprecated
+Deprecated functionality and modules. This is just a temporary holder before delete.
+Namely to look things up and port functionality if needed.
+
+You should not import from this.
