+++ tellurium/dev/oven/README.md
@@ -1,3 +1,3 @@
-# Oven
+# oven
 In the oven are test scripts and small code pieces not ready for the main code.
-Just ignore the `oven` and do not import anything from it.
+Just ignore the `oven` and DO NOT IMPORT anything from it.
