+++ tellurium/utils/matrix.py
@@ -1,84 +1,9 @@
-from __future__ import print_function, division, absolute_import
-import os
-import sys
-import warnings
-import functools
+"""
+Helpers for matrix operations.
+"""
 
-
-# ---------------------------------------------------------------------
-# Simple File Read and Store Utilities
-# ---------------------------------------------------------------------
-def saveToFile(filePath, str):
-    """ Save string to file.
-
-    see also: :func:`readFromFile`
-
-    :param filePath: file path to save to
-    :param str: string to save
-    """
-    with open(filePath, 'w') as f:
-        f.write(str)
-
-
-def readFromFile(filePath):
-    """ Load a file and return contents as a string.
-
-    see also: :func:`saveToFile`
-
-    :param filePath: file path to read from
-    :returns: string representation of the contents of the file
-    """
-    with open(filePath, 'r') as f:
-        string = f.read()
-    return string
-
-
-# ---------------------------------------------------------------------
-# Deprecated warning
-# ---------------------------------------------------------------------
-def deprecated(func):
-    """This is a decorator which can be used to mark functions
-        as deprecated. It will result in a warning being emitted
-        when the function is used.
-    """
-    @functools.wraps(func)
-    def new_func(*args, **kwargs):
-        warnings.warn_explicit(
-            "Call to deprecated function {}.".format(func.__name__),
-            category=DeprecationWarning,
-            filename=func.func_code.co_filename,
-            lineno=func.func_code.co_firstlineno + 1
-        )
-        return func(*args, **kwargs)
-    return new_func
-
-
-
-def runTool (toolFileName):
-    """ Call an external application called toolFileName.
-        Note that .exe extension may be omitted for windows applications.
-
-        Include any arguments in arguments parameter.
-
-        Example:
-        returnString = te.runTool (['myplugin', 'arg1', 'arg2'])
-
-              If the external tool writes to stdout, this will be captured and returned.
-
-        :param arguments to external tool
-        :return String return by external tool, if any.
-        """
-    import subprocess
-    try:
-        p = os.path.dirname(sys.executable)
-        root, waste = os.path.split (p)
-        #if (os.name == 'nt') and (not toolFileName[0].endswith ('.exe')):
-        toolFileName[0] = root + '\\telluriumTools\\' + toolFileName[0] + '\\' + toolFileName[0] + '.exe'
-        return subprocess.check_output(toolFileName)
-    except subprocess.CalledProcessError as e:
-        raise Exception ('Tool failed to run correctly or could not be found')
-
-
+from __future__ import absolute_import, printfunction
+import numpy as np
 
 def rank(A, atol=1e-13, rtol=0):
     """Estimate the rank (i.e. the dimension of the nullspace) of a matrix.
@@ -116,10 +41,11 @@ def rank(A, atol=1e-13, rtol=0):
     """
 
     A = np.atleast_2d(A)
-    s = svd(A, compute_uv=False)
+    s = np.linalg.svd(A, compute_uv=False)
     tol = max(atol, rtol * s[0])
     rank = int((s >= tol).sum())
     return rank
+
 
 def nullspace(A, atol=1e-13, rtol=0):
     """Compute an approximate basis for the nullspace of A.
@@ -155,20 +81,21 @@ def nullspace(A, atol=1e-13, rtol=0):
     """
 
     A = np.atleast_2d(A)
-    u, s, vh = svd(A)
+    u, s, vh = np.svd(A)
     tol = max(atol, rtol * s[0])
     nnz = (s >= tol).sum()
     ns = vh[nnz:].conj().T
     return ns
 
 
-# We import sympy here because it is slow to load and would slow down the initial
-# start up of tellurium
-def rref (A):
+def rref(A):
     """Compute the reduced row echelon for the matrix A. Returns
     returns a tuple of two elements. The first is the reduced row
     echelon form, and the second is a list of indices of the pivot columns.
     """
+
+    # We import sympy here because it is slow to load and would slow down the initial
+    # start up of tellurium
     import sympy
-    m = sympy.Matrix (A)
+    m = sympy.Matrix(A)
     return m.rref()
